Make prime generator return unsigned ints

+++ src/common/utils/FeedbackDelayNetwork.cpp
@@ -4,7 +4,7 @@
 namespace {
 auto generatePrimes(size_t count, unsigned startWith = 2)
 {
-  std::vector<int> primes;
+  std::vector<unsigned> primes;
   primes.reserve(count);
 
   const auto isPrime = [](unsigned n) {
@@ -19,7 +19,7 @@ auto generatePrimes(size_t count, unsigned startWith = 2)
   auto i = (startWith % 2 == 0) ? startWith + 1 : startWith;
   while (primes.size() < count) {
     if (isPrime(i))
-      primes.push_back(static_cast<int>(i));
+      primes.push_back(i);
     i += 2;
   }
